requirements.txt
django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-admin-thumbnails==0.2.9
django-filter==24.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1